+++ app/src/main/java/com/danosoftware/galaxyforce/sprites/game/aliens/implementations/StaticAlien.java
@@ -9,7 +9,6 @@ import com.danosoftware.galaxyforce.sprites.game.behaviours.powerup.PowerUpBehav
 import com.danosoftware.galaxyforce.sprites.game.behaviours.spawn.SpawnBehaviourFactory;
 import com.danosoftware.galaxyforce.sprites.game.behaviours.spinner.SpinningBehaviourFactory;
 import com.danosoftware.galaxyforce.waves.config.aliens.types.StaticConfig;
-
 import lombok.Builder;
 import lombok.NonNull;
 
@@ -20,35 +19,35 @@ public class StaticAlien extends AbstractAlien {
 
     @Builder
     public StaticAlien(
-            @NonNull final ExplosionBehaviourFactory explosionFactory,
-            @NonNull final SpawnBehaviourFactory spawnFactory,
-            @NonNull final SpinningBehaviourFactory spinningFactory,
-            @NonNull final PowerUpBehaviourFactory powerUpFactory,
-            @NonNull final FireBehaviourFactory fireFactory,
-            @NonNull final HitBehaviourFactory hitFactory,
-            @NonNull final StaticConfig alienConfig,
-            final PowerUpType powerUpType,
-            @NonNull final Integer xStart,
-            @NonNull final Integer yStart) {
+        @NonNull final ExplosionBehaviourFactory explosionFactory,
+        @NonNull final SpawnBehaviourFactory spawnFactory,
+        @NonNull final SpinningBehaviourFactory spinningFactory,
+        @NonNull final PowerUpBehaviourFactory powerUpFactory,
+        @NonNull final FireBehaviourFactory fireFactory,
+        @NonNull final HitBehaviourFactory hitFactory,
+        @NonNull final StaticConfig alienConfig,
+        final PowerUpType powerUpType,
+        @NonNull final Float xStart,
+        @NonNull final Float yStart) {
 
-        super(
-                alienConfig.getAlienCharacter(),
-                alienConfig.getAlienCharacter().getAnimation(),
-                xStart,
-                yStart,
-                alienConfig.getEnergy(),
-                fireFactory.createFireBehaviour(
-                        alienConfig.getMissileConfig()),
-                powerUpFactory.createPowerUpBehaviour(
-                        powerUpType),
-                spawnFactory.createSpawnBehaviour(
-                        alienConfig.getSpawnConfig()),
-                hitFactory.createHitBehaviour(
-                        alienConfig.getAlienCharacter().getHitAnimation()),
-                explosionFactory.createExplosionBehaviour(
-                        alienConfig.getExplosionConfig(),
-                        alienConfig.getAlienCharacter()),
-                spinningFactory.createSpinningBehaviour(
-                        alienConfig.getSpinningConfig()));
+      super(
+          alienConfig.getAlienCharacter(),
+          alienConfig.getAlienCharacter().getAnimation(),
+          xStart,
+          yStart,
+          alienConfig.getEnergy(),
+          fireFactory.createFireBehaviour(
+              alienConfig.getMissileConfig()),
+          powerUpFactory.createPowerUpBehaviour(
+              powerUpType),
+          spawnFactory.createSpawnBehaviour(
+              alienConfig.getSpawnConfig()),
+          hitFactory.createHitBehaviour(
+              alienConfig.getAlienCharacter().getHitAnimation()),
+          explosionFactory.createExplosionBehaviour(
+              alienConfig.getExplosionConfig(),
+              alienConfig.getAlienCharacter()),
+          spinningFactory.createSpinningBehaviour(
+              alienConfig.getSpinningConfig()));
     }
 }
